refactor(state): keep schema maintenance in its owner
Move the schema transaction guard, metadata writer, and canonical executor into the existing maintenance module. Preserve timestamp capture and synchronous transaction boundaries while keeping the runner within its existing line limit.
